Refactor changes related to streaming testcases
