--- tests/ballerina-unit-test/src/test/java/org/ballerinalang/test/streaming/ProjectionTest.java
@@ -1,72 +1,0 @@
-/*
-*  Copyright (c) 2018, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
-*
-*  WSO2 Inc. licenses this file to you under the Apache License,
-*  Version 2.0 (the "License"); you may not use this file except
-*  in compliance with the License.
-*  You may obtain a copy of the License at
-*
-*    http://www.apache.org/licenses/LICENSE-2.0
-*
-*  Unless required by applicable law or agreed to in writing,
-*  software distributed under the License is distributed on an
-*  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-*  KIND, either express or implied.  See the License for the
-*  specific language governing permissions and limitations
-*  under the License.
-*/
-package org.ballerinalang.test.streaming;
-
-import org.ballerinalang.launcher.util.BAssertUtil;
-import org.ballerinalang.launcher.util.BCompileUtil;
-import org.ballerinalang.launcher.util.BRunUtil;
-import org.ballerinalang.launcher.util.CompileResult;
-import org.ballerinalang.model.values.BInteger;
-import org.ballerinalang.model.values.BMap;
-import org.ballerinalang.model.values.BValue;
-import org.testng.Assert;
-import org.testng.annotations.BeforeClass;
-import org.testng.annotations.Test;
-
-/**
- * This contains methods to test projection behaviour of Ballerina Streaming.
- *
- * @since 0.965.0
- */
-public class ProjectionTest {
-
-    private CompileResult result;
-    private CompileResult resultNegative;
-
-    @BeforeClass
-    public void setup() {
-        result = BCompileUtil.compile("test-src/streaming/projection-streaming-test.bal");
-        resultNegative = BCompileUtil.compile("test-src/streaming/projection-streaming-negative-test.bal");
-    }
-
-    @Test(description = "Test projection streaming query")
-    public void testProjectionQuery() {
-        BValue[] outputEmployeeEvents = BRunUtil.invoke(result, "startProjectionQuery");
-
-        Assert.assertNotNull(outputEmployeeEvents);
-
-        Assert.assertEquals(outputEmployeeEvents.length, 3, "Expected events are not received");
-
-        BMap<String, BValue> employee0 = (BMap<String, BValue>) outputEmployeeEvents[0];
-        BMap<String, BValue> employee1 = (BMap<String, BValue>) outputEmployeeEvents[1];
-        BMap<String, BValue> employee2 = (BMap<String, BValue>) outputEmployeeEvents[2];
-
-        Assert.assertEquals(employee0.get("name").stringValue(), "Raja");
-        Assert.assertEquals(((BInteger) employee1.get("age")).intValue(), 33);
-        Assert.assertEquals(employee2.get("status").stringValue(), "married");
-    }
-
-    @Test(description = "Test streaming projection query with errors")
-    public void testProjectionNegativeCases() {
-        Assert.assertEquals(resultNegative.getErrorCount(), 1);
-        BAssertUtil.validateError(resultNegative, 0,
-                "incompatible stream action argument type 'Employee' defined",
-                45, 9);
-    }
-
-}
