+++ tests/ballerina-unit-test/src/test/java/org/ballerinalang/test/streaming/legacy/BooleanTypeTest.java
@@ -15,11 +15,12 @@
  *  specific language governing permissions and limitations
  *  under the License.
  */
-package org.ballerinalang.test.streaming;
+package org.ballerinalang.test.streaming.legacy;
 
 import org.ballerinalang.launcher.util.BCompileUtil;
 import org.ballerinalang.launcher.util.BRunUtil;
 import org.ballerinalang.launcher.util.CompileResult;
+import org.ballerinalang.model.values.BBoolean;
 import org.ballerinalang.model.values.BInteger;
 import org.ballerinalang.model.values.BMap;
 import org.ballerinalang.model.values.BValue;
@@ -28,30 +29,29 @@ import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 
 /**
- * This contains methods to test stream functions' behaviour of Ballerina Streaming.
+ * This contains methods to test boolean type conversion of Ballerina Streaming.
  *
- * @since 0.970.0
+ * @since 0.970.1
  */
-public class StreamFunctionTest {
+public class BooleanTypeTest {
 
     private CompileResult result;
 
     @BeforeClass
     public void setup() {
-        result = BCompileUtil.compile("test-src/streaming/stream-functions-test.bal");
+        result = BCompileUtil.compile("test-src/streaming/legacy/boolean-type-test.bal");
     }
 
-    @Test(description = "Test stream functions")
+    @Test(description = "Test Boolean types in streaming query")
     public void testFilterQuery() {
-        BValue[] outputEmployeeEvents = BRunUtil.invoke(result, "startStreamingQuery");
-        Assert.assertNotNull(outputEmployeeEvents);
+        BValue[] outputRequestCountObjs = BRunUtil.invoke(result, "startStreamingQuery");
+        Assert.assertNotNull(outputRequestCountObjs);
 
-        Assert.assertEquals(outputEmployeeEvents.length, 2, "Expected events are not received");
+        Assert.assertEquals(outputRequestCountObjs.length, 1, "Expected events are not received");
 
-        BMap<String, BValue> employee0 = (BMap<String, BValue>) outputEmployeeEvents[0];
-        BMap<String, BValue> employee1 = (BMap<String, BValue>) outputEmployeeEvents[1];
+        BMap<String, BValue> requestCount = (BMap<String, BValue>) outputRequestCountObjs[0];
 
-        Assert.assertEquals(((BInteger) employee0.get("age")).intValue(), 33);
-        Assert.assertEquals(((BInteger) employee1.get("age")).intValue(), 45);
+        Assert.assertEquals(((BInteger) requestCount.get("count")).intValue(), 7);
+        Assert.assertTrue(((BBoolean) requestCount.get("test")).booleanValue());
     }
 }
