+++ tests/ballerina-unit-test/src/test/java/org/ballerinalang/test/streaming/legacy/OrderByTest.java
@@ -15,7 +15,7 @@
  *  specific language governing permissions and limitations
  *  under the License.
  */
-package org.ballerinalang.test.streaming;
+package org.ballerinalang.test.streaming.legacy;
 
 import org.ballerinalang.launcher.util.BCompileUtil;
 import org.ballerinalang.launcher.util.BRunUtil;
@@ -38,7 +38,7 @@ public class OrderByTest {
 
     @BeforeClass
     public void setup() {
-        result = BCompileUtil.compile("test-src/streaming/orderby-streaming-test.bal");
+        result = BCompileUtil.compile("test-src/streaming/legacy/orderby-streaming-test.bal");
     }
 
     @Test(description = "Test order by in streaming query - a simple one")
