+++ tests/ballerina-unit-test/src/test/java/org/ballerinalang/test/streaming/legacy/PassthroughTest.java
@@ -15,11 +15,12 @@
 *  specific language governing permissions and limitations
 *  under the License.
 */
-package org.ballerinalang.test.streaming;
+package org.ballerinalang.test.streaming.legacy;
 
 import org.ballerinalang.launcher.util.BCompileUtil;
 import org.ballerinalang.launcher.util.BRunUtil;
 import org.ballerinalang.launcher.util.CompileResult;
+import org.ballerinalang.model.values.BInteger;
 import org.ballerinalang.model.values.BMap;
 import org.ballerinalang.model.values.BValue;
 import org.testng.Assert;
@@ -27,42 +28,53 @@ import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 
 /**
- * This contains methods to test output rate limiting behaviour of Ballerina Streaming.
+ * This contains methods to test passthrough behaviour of Ballerina Streaming.
  *
  * @since 0.965.0
  */
-public class OutputRateLimitTest {
+public class PassthroughTest {
 
     private CompileResult result;
-    private CompileResult resultForTimeBasedRateLimiting;
+    private CompileResult result2;
 
     @BeforeClass
     public void setup() {
-        result = BCompileUtil.compile("test-src/streaming/output-rate-limiting-test.bal");
-        resultForTimeBasedRateLimiting = BCompileUtil.compile("test-src/streaming/output-rate-limiting-time-test.bal");
+        result = BCompileUtil.compile("test-src/streaming/legacy/passthrough-streaming-test.bal");
+        result2 = BCompileUtil.compile("test-src/streaming/legacy/passthrough-streaming-without-select-test.bal");
     }
 
-    @Test(description = "Test output rate limiting query")
-    public void testOutputRateLimitQuery() {
-        BValue[] outputEmployeeEvents = BRunUtil.invoke(result, "startOutputRateLimitQuery");
+    @Test(description = "Test passthrough streaming query")
+    public void testPassthroughQuery() {
+        BValue[] outputEmployeeEvents = BRunUtil.invoke(result, "startPassthroughQuery");
+
         Assert.assertNotNull(outputEmployeeEvents);
 
-        Assert.assertEquals(outputEmployeeEvents.length, 2, "Expected events are not received");
+        Assert.assertEquals(outputEmployeeEvents.length, 3, "Expected events are not received");
 
         BMap<String, BValue> employee0 = (BMap<String, BValue>) outputEmployeeEvents[0];
         BMap<String, BValue> employee1 = (BMap<String, BValue>) outputEmployeeEvents[1];
+        BMap<String, BValue> employee2 = (BMap<String, BValue>) outputEmployeeEvents[2];
 
         Assert.assertEquals(employee0.get("name").stringValue(), "Raja");
-        Assert.assertEquals(employee1.get("name").stringValue(), "Praveen");
+        Assert.assertEquals(((BInteger) employee1.get("age")).intValue(), 33);
+        Assert.assertEquals(employee2.get("status").stringValue(), "married");
     }
 
-    @Test(description = "Test output rate limiting query based on time")
-    public void testOutputRateLimitWithTimeQuery() {
-        BValue[] outputEmployeeEvents = BRunUtil.invoke(resultForTimeBasedRateLimiting, "startOutputRateLimitQuery");
+    @Test(description = "Test passthrough streaming query without select")
+    public void testPassthroughQueryWithoutSelect() {
+        BValue[] outputEmployeeEvents = BRunUtil.invoke(result, "startPassthroughQuery");
+
         Assert.assertNotNull(outputEmployeeEvents);
-        Assert.assertEquals(outputEmployeeEvents.length, 1, "Expected events are not received");
+
+        Assert.assertEquals(outputEmployeeEvents.length, 3, "Expected events are not received");
 
         BMap<String, BValue> employee0 = (BMap<String, BValue>) outputEmployeeEvents[0];
+        BMap<String, BValue> employee1 = (BMap<String, BValue>) outputEmployeeEvents[1];
+        BMap<String, BValue> employee2 = (BMap<String, BValue>) outputEmployeeEvents[2];
+
         Assert.assertEquals(employee0.get("name").stringValue(), "Raja");
+        Assert.assertEquals(((BInteger) employee1.get("age")).intValue(), 33);
+        Assert.assertEquals(employee2.get("status").stringValue(), "married");
     }
+
 }
