+++ tests/ballerina-unit-test/src/test/java/org/ballerinalang/test/streaming/legacy/ProjectionTest.java
@@ -15,8 +15,9 @@
 *  specific language governing permissions and limitations
 *  under the License.
 */
-package org.ballerinalang.test.streaming;
+package org.ballerinalang.test.streaming.legacy;
 
+import org.ballerinalang.launcher.util.BAssertUtil;
 import org.ballerinalang.launcher.util.BCompileUtil;
 import org.ballerinalang.launcher.util.BRunUtil;
 import org.ballerinalang.launcher.util.CompileResult;
@@ -28,31 +29,44 @@ import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 
 /**
- * This contains methods to test filter behaviour of Ballerina Streaming.
+ * This contains methods to test projection behaviour of Ballerina Streaming.
  *
  * @since 0.965.0
  */
-public class FilterWithInlineStreamTest {
+public class ProjectionTest {
 
     private CompileResult result;
+    private CompileResult resultNegative;
 
     @BeforeClass
     public void setup() {
-        result = BCompileUtil.compile("test-src/streaming/filter-streaming-test-v2.bal");
+        result = BCompileUtil.compile("test-src/streaming/legacy/projection-streaming-test.bal");
+        resultNegative = BCompileUtil.compile("test-src/streaming/negative/projection-streaming-negative-test.bal");
     }
 
-    @Test(description = "Test filter streaming query")
-    public void testFilterQuery() {
-        BValue[] outputEmployeeEvents = BRunUtil.invoke(result, "startFilterQuery");
+    @Test(description = "Test projection streaming query")
+    public void testProjectionQuery() {
+        BValue[] outputEmployeeEvents = BRunUtil.invoke(result, "startProjectionQuery");
+
         Assert.assertNotNull(outputEmployeeEvents);
 
-        Assert.assertEquals(outputEmployeeEvents.length, 2, "Expected events are not received");
+        Assert.assertEquals(outputEmployeeEvents.length, 3, "Expected events are not received");
 
         BMap<String, BValue> employee0 = (BMap<String, BValue>) outputEmployeeEvents[0];
         BMap<String, BValue> employee1 = (BMap<String, BValue>) outputEmployeeEvents[1];
+        BMap<String, BValue> employee2 = (BMap<String, BValue>) outputEmployeeEvents[2];
 
-        Assert.assertEquals(((BInteger) employee0.get("age")).intValue(), 33);
-        Assert.assertEquals(((BInteger) employee1.get("age")).intValue(), 45);
+        Assert.assertEquals(employee0.get("name").stringValue(), "Raja");
+        Assert.assertEquals(((BInteger) employee1.get("age")).intValue(), 33);
+        Assert.assertEquals(employee2.get("status").stringValue(), "married");
+    }
+
+    @Test(description = "Test streaming projection query with errors")
+    public void testProjectionNegativeCases() {
+        Assert.assertEquals(resultNegative.getErrorCount(), 1);
+        BAssertUtil.validateError(resultNegative, 0,
+                "incompatible stream action argument type 'Employee' defined",
+                45, 9);
     }
 
 }
