+++ tests/ballerina-unit-test/src/test/resources/test-src/streaming/legacy/aggregation-streaming-test.bal
