+++ tests/ballerina-unit-test/src/test/resources/test-src/streaming/legacy/boolean-type-test.bal
