+++ tests/ballerina-unit-test/src/test/resources/test-src/streaming/legacy/filter-streaming-with-reference-test.bal
