+++ tests/ballerina-unit-test/src/test/resources/test-src/streaming/legacy/join-streaming-test.bal
