+++ tests/ballerina-unit-test/src/test/resources/test-src/streaming/legacy/output-rate-limiting-test.bal
