+++ tests/ballerina-unit-test/src/test/resources/test-src/streaming/legacy/passthrough-streaming-without-select-test.bal
