+++ tests/ballerina-unit-test/src/test/resources/test-src/streaming/legacy/pattern-streaming-test.bal
