+++ tests/ballerina-unit-test/src/test/resources/test-src/streaming/legacy/pipeline-streaming-test.bal
