+++ tests/ballerina-unit-test/src/test/resources/test-src/streaming/legacy/projection-streaming-test.bal
