+++ tests/ballerina-unit-test/src/test/resources/test-src/streaming/legacy/sequence-streaming-test.bal
