+++ tests/ballerina-unit-test/src/test/resources/test-src/streaming/legacy/stream-functions-test.bal
