+++ tests/ballerina-unit-test/src/test/resources/test-src/streaming/legacy/streaming-inline-operations-test.bal
