+++ tests/ballerina-unit-test/src/test/resources/test-src/streaming/negative/projection-streaming-negative-test.bal
