+++ tests/ballerina-unit-test/src/test/resources/test-src/streaming/negative/streaming-invalid-record-type-negative-test.bal
